[FEAT] canvasList socket 재설정
